requirements.txt
# Core dependencies
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0

# Transfer methods
paramiko==3.3.1  # For SFTP transfers

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development tools (optional)
black==23.11.0
flake8==6.1.0
isort==5.12.0
